Extract InternalDateTime to a new header file
Also some resolving of minor warnings by Clang-Tidy like "explicit" and "override" specifiers.
